# 🦟 Bug fix


## Summary

Alternative to https://github.com/gazebosim/gz-sim/pull/3070. Instead of completely skipping the model serialization / materialization step, this PR uses a workaround to speed up the deserialization process. This is done by implementing a modified version of [sdf::Root::LoadSdfString](https://github.com/gazebosim/sdformat/blob/067c73bc5328c9687da1b0e43588f324722e5e7a/src/Root.cc#L217). It uses a static `sdf::SDFPtr` object to avoid multiple expensive `sdf::init` calls. Note that this workaround is similar to the one done in [sdf/src/pasrer.cc](https://github.com/gazebosim/sdformat/blob/067c73bc5328c9687da1b0e43588f324722e5e7a/src/parser.cc#L1741-L1745). 

We could also implement this workaround directly in sdformat but this I chose to implement it here to reduce the impact of this change during the code freeze period.

For the [Jetty demo world ](https://github.com/gazebosim/jetty_demo), the load times are:

* Before any changes: 55s
* Skip model sdf serialization (#3070): 42s
* this pull request: 48s

## Checklist
- [x] Signed all commits for DCO
- [ ] Added tests
- [ ] Updated documentation (as needed)
- [ ] Updated migration guide (as needed)
- [ ] Consider updating Python bindings (if the library has them)
- [ ] `codecheck` passed (See [contributing](https://gazebosim.org/docs/all/contributing#contributing-code))
- [ ] All tests passed (See [test coverage](https://gazebosim.org/docs/all/contributing#test-coverage))
- [ ] While waiting for a review on your PR, please help review [another open pull request](https://github.com/pulls?q=is%3Aopen+is%3Apr+user%3Agazebosim+archived%3Afalse+) to support the maintainers
- [ ] Was GenAI used to generate this PR? If so, make sure to add "Generated-by" to your commits. (See [this policy](https://osralliance.org/wp-content/uploads/2025/05/OSRF-Policy-on-the-Use-of-Generative-Tools-Generative-AI-in-Contributions.pdf) for more info.)

Generated-by: Remove this if GenAI was not used.

**Note to maintainers**: Remember to use **Squash-Merge** and edit the commit message to match the pull request summary while retaining `Signed-off-by` and `Generated-by` messages.