@iche033 2025-09-24
Testing this with the Jetty demo world and https://github.com/gazebosim/sdformat/pull/1589,  I see that this PR still improves performance.

The initial load time dropped from 33 seconds to 22 seconds on my machine, and resuming to play state from paused state reduced from 17 to 3 seconds.