@arjo129 2025-10-01
Thanks Ian! Just to clarify Im assuming we are sure that serialization and deserialization only ever happen in a single threaded context